requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
Django==5.1.2
djangorestframework==3.15.2
future==1.0.0
geocoder==1.38.1
idna==3.10
mysqlclient==2.2.5
pillow==11.0.0
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
